+++ CHANGES.rst
@@ -5,7 +5,7 @@
 5.0 (unreleased)
 ================
 
-- Nothing changed yet.
+- Drop support for Python 2.7, 3.5, 3.6.
 
 
 4.2.1 (2022-08-25)